# services/api-gateway/api.yaml
openapi: 3.1.0
info:
  title: JIMS-AI api-gateway
  version: 0.1.0
paths:
...
  /v1/training/panels/{panel}/items:
    get:
      summary: Cursor-paged stored data for a training panel
      parameters:
        - name: panel
          in: path
          required: true
          schema:
            type: string
            enum: [ingestion, review, ambiguity, memory, world-model, pipeline, sessions, feedback]
        - name: cursor
          in: query
          required: false
          schema:
            type: string
        - name: limit
          in: query
          required: false
          schema:
            type: integer
            default: 25
      responses:
        "200":
          description: Paged panel records
...
  /v1/training/kaggle/{run_id}/sync:
    post:
      summary: Check Kaggle run status and download completed artifacts
      parameters:
        - name: run_id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Kaggle run status and artifact state